ImmigrationAI/api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
openai>=1.52.0
httpx==0.27.0
qdrant-client==1.9.1
python-dotenv==1.0.1
pydantic==2.7.1